requirements.txt
openai
beautifulsoup4
requests
Pillow
python-dotenv